implement server

+++ pkg/model/edge.go
@@ -1,9 +1,9 @@
 package model
 
 type Edge struct {
-	Id          int     `json:"id" db:"id"`
-	StartRoomId int     `json:"start_room_id" db:"start_room_id"`
-	EndRoomId   int     `json:"end_room_id" db:"end_room_id"`
+	Id          int64   `json:"id" db:"id"`
+	StartRoomId int64   `json:"start_room_id" db:"start_room_id"`
+	EndRoomId   int64   `json:"end_room_id" db:"end_room_id"`
 	Cost        float64 `json:"cost" db:"cost"`
-	CampusId    int     `json:"campus_id" db:"campus_id"`
+	CampusId    int64   `json:"campus_id" db:"campus_id"`
 }
