+++ pkg/model/room.go
@@ -1,7 +1,7 @@
 package model
 
 type Room struct {
-	Id       int    `json:"id" db:"id"`
+	Id       int64  `json:"id" db:"id"`
 	Code     string `json:"code" db:"code"`
-	CampusId int    `json:"campus_id" db:"campus_id"`
+	CampusId int64  `json:"campus_id" db:"campus_id"`
 }
