+++ pkg/model/time.go
@@ -1,8 +1,8 @@
 package model
 
 type Time struct {
-	Id     int `json:"id" db:"id"`
-	GameId int `json:"game_id" db:"game_id"`
-	UserId int `json:"user_id" db:"user_id"`
-	Time   int `json:"time" db:"time"`
+	Id     int64 `json:"id" db:"id"`
+	GameId int64 `json:"game_id" db:"game_id"`
+	UserId int64 `json:"user_id" db:"user_id"`
+	Time   int64 `json:"time" db:"time"`
 }
