+++ pkg/repository/game_postgres.go
@@ -11,7 +11,7 @@ type GamePostgres struct {
 	db *sqlx.DB
 }
 
-func (g *GamePostgres) GetGame(gameId int) (model.GameUsers, error) {
+func (g *GamePostgres) GetGame(gameId int64) (model.GameUsers, error) {
 	var game model.GameUsers
 	query := fmt.Sprintf("SELECT * FROM %s WHERE id=$1", gamesTable)
 	err := g.db.Get(&game, query, gameId)
@@ -19,7 +19,7 @@ func (g *GamePostgres) GetGame(gameId int) (model.GameUsers, error) {
 	return game, err
 }
 
-func (g *GamePostgres) GetTime(gameId, userId int) (model.Time, error) {
+func (g *GamePostgres) GetTime(gameId, userId int64) (model.Time, error) {
 	var time model.Time
 	query := fmt.Sprintf("SELECT * FROM %s WHERE game_id=$1 AND user_id=$2", timesTable)
 	err := g.db.Get(&time, query, gameId, userId)
@@ -27,8 +27,8 @@ func (g *GamePostgres) GetTime(gameId, userId int) (model.Time, error) {
 	return time, err
 }
 
-func (g *GamePostgres) AddGame(userIdFirst, userIdSecond int) (int, error) {
-	var id int
+func (g *GamePostgres) AddGame(userIdFirst, userIdSecond int64) (int64, error) {
+	var id int64
 	query := fmt.Sprintf("INSERT INTO %s (user_id_first, user_id_second) values ($1, $2) RETURNING id", gamesTable)
 	row := g.db.QueryRow(query, userIdFirst, userIdSecond)
 	if err := row.Scan(&id); err != nil {
@@ -38,21 +38,21 @@ func (g *GamePostgres) AddGame(userIdFirst, userIdSecond int) (int, error) {
 	return id, nil
 }
 
-func (g *GamePostgres) AddTime(gameId, userId, time int) error {
+func (g *GamePostgres) AddTime(gameId, userId, time int64) error {
 	query := fmt.Sprintf("INSERT INTO %s (game_id, user_id, time) values ($1, $2, $3)", timesTable)
 	_, err := g.db.Exec(query, gameId, userId, time)
 
 	return err
 }
 
-func (g *GamePostgres) UpdateTime(gameId, userId, time int) error {
+func (g *GamePostgres) UpdateTime(gameId, userId, time int64) error {
 	query := fmt.Sprintf("UPDATE %s SET time = $1 WHERE game_id = $2 AND user_id = $3", timesTable)
 	_, err := g.db.Exec(query, time, gameId, userId)
 
 	return err
 }
 
-func (g *GamePostgres) GetRoomById(roomId int) (model.Room, error) {
+func (g *GamePostgres) GetRoomById(roomId int64) (model.Room, error) {
 	var room model.Room
 	query := fmt.Sprintf(`SELECT * FROM %s rm WHERE rm.id = $1`, roomsTable)
 	err := g.db.Get(&room, query, roomId)
@@ -60,8 +60,8 @@ func (g *GamePostgres) GetRoomById(roomId int) (model.Room, error) {
 	return room, err
 }
 
-func (g *GamePostgres) AddCall(userIdFirst, userIdSecond, roomIdFirst int) (model.Game, error) {
-	var count int
+func (g *GamePostgres) AddCall(userIdFirst, userIdSecond, roomIdFirst int64) (model.Game, error) {
+	var count int64
 	tx, err := g.db.Begin()
 	if err != nil {
 		return model.Game{}, err
@@ -119,13 +119,13 @@ func (g *GamePostgres) AddCall(userIdFirst, userIdSecond, roomIdFirst int) (mode
 	}
 }
 
-func (g *GamePostgres) DeleteCall(userIdFirst, userIdSecond int) error {
+func (g *GamePostgres) DeleteCall(userIdFirst, userIdSecond int64) error {
 	query := fmt.Sprintf("DELETE FROM %s WHERE user_id_first=$1 AND user_id_second=$2", callsTable)
 	_, err := g.db.Exec(query, userIdFirst, userIdSecond)
 	return err
 }
 
-func (g *GamePostgres) GetListOfEdges(startRoomId int) ([]model.Edge, error) {
+func (g *GamePostgres) GetListOfEdges(startRoomId int64) ([]model.Edge, error) {
 	var edges []model.Edge
 
 	query := fmt.Sprintf(`SELECT * FROM %s ed WHERE ed.start_room_id = $1`, edgesTable)
@@ -134,7 +134,7 @@ func (g *GamePostgres) GetListOfEdges(startRoomId int) ([]model.Edge, error) {
 	return edges, err
 }
 
-func (g *GamePostgres) GetEdge(startRoomId, endRoomId int) (model.Edge, error) {
+func (g *GamePostgres) GetEdge(startRoomId, endRoomId int64) (model.Edge, error) {
 	var edges []model.Edge
 
 	query := fmt.Sprintf(`SELECT * FROM %s ed WHERE ed.start_room_id = $1 AND ed.end_room_id = $2`, edgesTable)
@@ -148,7 +148,7 @@ func (g *GamePostgres) GetEdge(startRoomId, endRoomId int) (model.Edge, error) {
 	return edges[0], err
 }
 
-func (g *GamePostgres) GetRoomByCodePattern(code string, campusId int) ([]model.Room, error) {
+func (g *GamePostgres) GetRoomByCodePattern(code string, campusId int64) ([]model.Room, error) {
 	var rooms []model.Room
 
 	query := fmt.Sprintf(`SELECT * FROM %s rm WHERE rm.campus_id = $1 AND rm.code LIKE $2`, roomsTable)
